website: proper meta tags
For more info and to test this, see [0].

It is better to use png by default and use a transparent webp
rewrite on the server, if the client accepts it.

[0] https://www.opengraph.xyz/url/https%3A%2F%2Fukvly.org%2F

meta: fix meta tag

This must start with a slash. I found out so by using [0].

[0] https://www.opengraph.xyz/url/https%3A%2F%2Fukvly.org%2F

meta: back to webp

meta: